% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_analysis.R
\name{area_analysis}
\alias{area_analysis}
\title{Analysis of monthly area averages
This function calculates area average for each time step, groups data by months and calculates statistics over time period of data set.
Calculated parameters are mean, median, min, max, Sen Slope, Mann-Kendall tau and Mann-Kendall p-value.}
\usage{
area_analysis(data)
}
\arguments{
\item{data}{Object of class SpatRaster.
Data can be monthly or monthly averaged by hour of day with one or more hours selected.}
}
\value{
A data frame with the following columns: months, mean, median, min, max, slope, mk_tau and mk_pval of each month over time for the area average of the data.
}
\description{
Analysis of monthly area averages
This function calculates area average for each time step, groups data by months and calculates statistics over time period of data set.
Calculated parameters are mean, median, min, max, Sen Slope, Mann-Kendall tau and Mann-Kendall p-value.
}
