% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_era5_land_request.R
\name{build_era5_land_request}
\alias{build_era5_land_request}
\title{Builds a request compatible with the CDS for all microclimate relevant climate
variables of ERA5-Land.}
\usage{
build_era5_land_request(
  xmin,
  xmax,
  ymin,
  ymax,
  start_time,
  end_time,
  outfile_name = "era5_land_out"
)
}
\arguments{
\item{xmin}{The minimum longitude to request data for.}

\item{xmax}{The maximum longitude to request data for.}

\item{ymin}{The minimum latitude to request data for.}

\item{ymax}{The maximum latitude to request data for.}

\item{start_time}{a POSIXlt object indicating the first hour for which data
are required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
are required.}

\item{outfile_name}{character prefix for .nc files when downloaded (the year,
month, and file extension is automatically added). Keep in mind that you may
want to submit multiple (or many) requests, and therefore this prefix should
adequately describe a unique query (e.g. by its spatial or temporal extent).}
}
\description{
`build_era5_land_request` creates a request or set of requests of
ERA5-Land data that can be submitted to the Climate Data Store (CDS) with
the `ecmwfr` package.
Spatial and temporal extents are defined by the user, and requests are
automatically split by month. The following variables are requested:
2m_temperature, 2m_dewpoint_temperature, surface_pressure,
10m_u_component_of_wind, 10m_v_component_of_wind, total_precipitation. Note that
these variables are not sufficient to run several microclimate models, e.g.
`microclimc` and `microclimf`.
}
