% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_netcdf.R
\name{combine_netcdf}
\alias{combine_netcdf}
\title{Combines a series of netCDFs that all have the same spatial extent,
set of variables, and units of time}
\usage{
combine_netcdf(filenames, combined_name)
}
\arguments{
\item{filenames}{a list of filenames for netCDFs you wish to combine}

\item{combined_name}{the name of the combined netCDF, must end with ".nc"}
}
\value{
No object is returned, the combined netCDF file is written to disk
with the name specified by `combined_name`
}
\description{
Combines a series of netCDFs that all have the same spatial extent,
set of variables, and units of time
}
