% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hours_to_seconds.R
\name{convert_hours_to_seconds}
\alias{convert_hours_to_seconds}
\title{Convert hours to seconds for the timeout parameter in the ERA5 function}
\usage{
convert_hours_to_seconds(time_out)
}
\arguments{
\item{time_out}{Character string indicating the time to wait for the data to be downloaded. \cr
It can be in hours (ending with 'h') or seconds (ending with 's'). \cr}
}
\value{
The time in seconds as input for the timeout parameter.
}
\description{
This function allows you to use hours or seconds as input for the timeout parameter in the \code{\link{get_era5}} function.
By default, you can only use seconds as input for the timeout parameter. This function allows you to use hours as input as well.
The timeout parameter is used to set the maximum time to wait for the data to be downloaded from the Copernicus Climate Data Store.
}
