% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Function to convert units to more common units}
\usage{
convert_units(data)
}
\arguments{
\item{data}{Object of class SpatRaster.}
}
\value{
An object of class SpatRaster with converted units if units are in the list of units to be converted
}
\description{
This function converts values and units from data sets which are in Kelvin (to °C), meters (to mm), meters of water equivalent (to mm of water equivalent), and Pascals (to hPa).
}
