% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clim.R
\name{extract_clim}
\alias{extract_clim}
\title{Produces hourly data for a single location ready for use with
`microclima::runauto`.}
\usage{
extract_clim(
  nc,
  long,
  lat,
  start_time,
  end_time,
  d_weight = TRUE,
  dtr_cor = TRUE,
  dtr_cor_fac = 1.285
)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
`build_era5_request` and `request_era5` functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long}{longitude of the location for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat}{latitude of the location for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt object indicating the first hour for which data
are required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
are required.}

\item{d_weight}{logical value indicating whether to apply inverse distance
weighting using the 4 closest neighbouring points to the location defined by
`long` and `lat`. Default = `TRUE`.}

\item{dtr_cor}{logical value indicating whether to apply a diurnal temperature
range correction to air temperature values. Default = `TRUE`.}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction. Default = 1.285, based on calibration against UK Met Office
observations.}
}
\value{
a data frame containing hourly values for a suite of climate variables:

`obs_time` | the date-time (timezone specified in col timezone)

`temperature` | (degrees celsius)

`humidity` | specific humidity (kg / kg)

`pressure` | (Pa)

`windspeed` | (m / s)

`winddir` | wind direction, azimuth (degrees from north)

`emissivity` | downward long wave radiation flux divided by the sum
of net long-wave radiation flux and downward long wave radiation flux (unitless)

`cloudcover` | (percent)

`netlong` | Net longwave radiation (MJ / m2 / hr)

`uplong` | Upward longwave radiation (MJ / m2 / hr)

`downlong` | Downward longwave radiation (MJ / m2 / hr)

`rad_dni` | Direct normal irradiance (MJ / m2 / hr)

`rad_dif` | Diffuse normal irradiance (MJ / m2 / hr)

`szenith` | Solar zenith angle (degrees from a horizontal plane)

`timezone` | (unitless)
}
\description{
`extract_clim` takes an nc file containing hourly ERA5 climate
data, and for a given set of coordinates, produces an (optionally) inverse
distance weighted mean of each variable, ready for use with
`microclima::runauto`. Also provides the option to implement a diurnal
temperature range correction to air temperatures.
}
