% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_date_components.R
\name{extract_date_components}
\alias{extract_date_components}
\title{Extracting date components for the desired format of the API}
\usage{
extract_date_components(
  StartDate,
  EndDate,
  dataset_short_name,
  year_interval = 1,
  month_interval = 1,
  day_interval = 1,
  target_years = NULL,
  target_days = NULL,
  season = NULL,
  target_months = NULL
)
}
\arguments{
\item{StartDate}{A character string or Date object specifying the start date of the data extraction in the format "YYYY-MM-DD" (e.g. "2010-01-01").}

\item{EndDate}{A character string or Date object specifying the end date of the data extraction in the format "YYYY-MM-DD" (e.g. "2018-11-05").}

\item{dataset_short_name}{A character string specifying the dataset short name (e.g. "ERA5-Land monthly averaged data") as in the API
It is used to check if the dataset is monthly or daily. If the dataset is monthly, days will be an empty list.}

\item{year_interval}{An integer specifying the interval for the years (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{month_interval}{An integer specifying the interval for the months (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{day_interval}{An integer specifying the interval for the days (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{target_years}{A character vector specifying the years to be selected (e.g. c("2015", "2016", "2017")). If provided, only these years will be selected and the months and days adjusted (optional parameter).}

\item{target_days}{A character vector specifying the days to be selected (e.g. c("15", "16", "17")). If provided, only these days will be selected and the months and years adjusted (optional parameter).\cr
If the dataset is monthly, it is not possible to select days, the parameter will be ignored and days will be empty.}

\item{season}{A character string specifying the season to be selected (e.g. "DJF", "MAM", "JJA", "SON") (optional parameter) \cr}

\item{target_months}{A character vector specifying the months to be selected (e.g. c("01", "02", "03")). If provided, only these months will be selected and the days and years adjusted (optional parameter).}
}
\value{
A list containing the extracted date components: \cr
\itemize{
\item year: A character vector containing the years \cr
\item month: A character vector containing the months \cr
\item day: A character vector containing the days
}
}
\description{
The Copernicus Climate Data Store (CDS) API requires the date components for downloading the data to be in a specific format. \cr
It requires one list for the years, one for the months and one for the days. \cr
This function allows the user to set a StartDate and EndDate, and the function will automatically extract the years, months, and days, if no other specifications are set. \cr
In addition, the user can filter the dates based on the following parameters: \cr
\itemize{
\item target_years: select specific years \cr
\item target_months: select specific months \cr
\item target_days: select specific days \cr
\item season: select a season \cr
\item year_interval: select every nth year \cr
\item month_interval: select every nth month \cr
\item day_interval: select every nth day \cr
}
}
