% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_precip.R
\name{extract_precip}
\alias{extract_precip}
\title{Produces daily or hourly precipitation data for a single location ready for use
with `microclima::runauto`.}
\usage{
extract_precip(
  nc,
  long,
  lat,
  start_time,
  end_time,
  d_weight = TRUE,
  convert_daily = TRUE,
  cds_version = "new"
)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
`build_era5_request` and `request_era5` functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long}{longitude of the location for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat}{latitude of the location for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt or POSIXct object indicating the first day or hour
for which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{end_time}{a POSIXlt or POSIXct object indicating the last day or hour for
which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{d_weight}{logical value indicating whether to apply inverse distance
weighting using the 4 closest neighbouring points to the location defined by
`long` and `lat`. Default = `TRUE`.}

\item{convert_daily}{a flag indicating whether the user desires to convert the
precipitation vector from hourly to daily averages (TRUE) or remain as hourly
values (FALSE). Only daily precipitation will be accepted by `microclima::runauto`.}

\item{cds_version}{specifies what version of the CDS (Climate Data Store) the
ERA5 data were queried from. Either "new" (queries made after Sept 2024) or
"legacy" (queries made before Sept 2024). This argument will be deprecated in
the future and only queries from the new CDS will be supported.}
}
\value{
a numeric vector of daily or hourly precipitation (mm).
}
\description{
`extract_precip` takes an nc file containing hourly ERA5
climate data, and for a given set of coordinates, produces an (optionally)
inverse distance weighted mean of precipitation (at daily or hourly resolution)
ready for use with `microclima::runauto`.
}
