% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_precipa.R
\name{extract_precipa}
\alias{extract_precipa}
\title{Produces daily or hourly precipitation data across a grid ready for use with
several gridded microclimate models.}
\usage{
extract_precipa(
  nc,
  long_min,
  long_max,
  lat_min,
  lat_max,
  start_time,
  end_time,
  convert_daily = TRUE,
  cds_version = "new"
)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
`build_era5_request` and `request_era5` functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long_min}{minimum longitude of the grid for which data are required (decimal}

\item{long_max}{maximum longitude of the grid for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat_min}{minimum latitude of the grid for which data are required (decimal}

\item{lat_max}{maximum latitude of the grid for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt or POSIXct object indicating the first day or hour
for which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{end_time}{a POSIXlt or POSIXct object indicating the last day or hour for
which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{convert_daily}{a flag indicating whether the user desires to convert the
 precipitation spatRaster from hourly to daily averages (TRUE) or remain as hourly
values (FALSE). Only daily precipitation will be accepted by `microclimf::modelina`.}

\item{cds_version}{specifies what version of the CDS (Climate Data Store) the
ERA5 data were queried from. Either "new" (queries made after Sept 2024) or
"legacy" (queries made before Sept 2024). This argument will be deprecated in
the future and only queries from the new CDS will be supported.}
}
\value{
a spatRaster of daily or hourly precipitation (mm).
}
\description{
`extract_precipa` takes an nc file containing hourly ERA5
climate data, and for a given set of coordinates, produces precipitation
(at daily or hourly resolution) ready for use with `microclimf::modelina`.
}
