% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_target_name.R
\name{generate_default_target_file_name}
\alias{generate_default_target_file_name}
\title{Generate default target name for the ncfile downloaded by the ERA5 function}
\usage{
generate_default_target_file_name(
  variable,
  StartDate,
  EndDate,
  target_file = NULL
)
}
\arguments{
\item{variable}{Character string. The variable name specified in the ERA5 function.
datatype:}

\item{StartDate}{Character string. The start date specified in the ERA5 function.}

\item{EndDate}{Character string. The end date specified in the ERA5 function.}

\item{target_file}{Character string. If the user has specified a name fo the ncfile, this will be used instead of the file name generated by this function (optional parameter)}
}
\value{
A character string for the name of the ncfile
}
\description{
This function generates a default target name for the ncfile downloaded by the ERA5 function. \cr
The default target name is generated by concatenating the variable name, StartDate and EndDate.
}
\examples{
variable <- "2m_temperature"
StartDate <- "2018-01-01"
EndDate <- "2018-01-31"
target_name <- generate_default_target_file_name(variable, StartDate, EndDate)
print(target_name)
"2m_temperature_2018-01-01__2018-01-31.nc"
}
