% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spatial_object_countries.R
\name{generate_spatial_obj}
\alias{generate_spatial_obj}
\title{Get the geometry of a selected country}
\usage{
generate_spatial_obj(country_name, admin_level = NULL)
}
\arguments{
\item{country_name}{A character. The name of the country for which the geometry is to be extracted.}

\item{admin_level}{Integer. The administrative level of the country. If not specified, the function will return the geometry of the country.}
}
\value{
A sfc Multipolygon containing the geometry of the selected country.
}
\description{
This function returns the geometry of a country chosen from the rgeoboundaries package  .\cr
The user can specify the country name and optionally the administrative level. \cr
For details view the documentation of the rgeoboundaries package (https://github.com/wmgeolab/rgeoboundaries).
}
