% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_ERA5.R
\name{get_era5}
\alias{get_era5}
\title{Download ERA5 data from the Copernicus Climate Data Store (CDS) API in R}
\usage{
get_era5(
  user,
  key,
  dataset_short_name,
  variable,
  StartDate,
  EndDate,
  time,
  aoi = NULL,
  time_out = NULL,
  year_interval = NULL,
  month_interval = NULL,
  day_interval = NULL,
  target_years = NULL,
  target_months = NULL,
  target_days = NULL,
  season = NULL,
  country_name = NULL,
  admin_level = NULL,
  pressure_level = NULL,
  target_file = NULL,
  all_hours = FALSE,
  path = NULL,
  product_type = NULL,
  grid = NULL
)
}
\arguments{
\item{user}{Character string. Your Copernicus Climate Data Store (CDS) API username}

\item{key}{Character string. Your Copernicus Climate Data Store (CDS) API key}

\item{dataset_short_name}{Character string.. The short name of the ERA5 dataset to download.}

\item{variable}{Character string. The variable to download from the ERA5 dataset.}

\item{StartDate}{A character string or Date object specifying the start date of the data extraction in the format "YYYY-MM-DD" (e.g. "2010-01-01").}

\item{EndDate}{A character string or Date object specifying the end date of the data extraction in the format "YYYY-MM-DD" (e.g. "2018-11-05").}

\item{time}{A character string or list of character strings specifying the time of the day to download the data (e.g.time <-sprintf("\%02d:00", 2:8)   \cr.
For monthly averaged datasets, the time is automatically set to '00:00'.}

\item{aoi}{The Area of interest (aoi) specified by the user. It can be any object of the sf or terra package.\cr
The function will automatically determine the extent of the aoi and round it to the grid resolution of the dataset. \cr
Instead, the user can also select coordinates in the format c(North/ West/South/East, e.g. 50.75/9/47.5/14).}

\item{time_out}{A character string specifying the time out for the download in hours or seconds. The default is 1 hour (optional parameter).}

\item{year_interval}{An integer specifying the interval for the years (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{month_interval}{An integer specifying the interval for the months (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{day_interval}{An integer specifying the interval for the days (e.g. 2,3,...). The default is 1 (optional parameter).}

\item{target_years}{A character vector specifying the years to be selected (e.g. c("2015", "2016", "2017")). If provided, only these years will be selected and the months and days adjusted (optional parameter).}

\item{target_months}{A character vector specifying the months to be selected (e.g. c("01", "02", "03")). If provided, only these months will be selected and the days and years adjusted (optional parameter).}

\item{target_days}{A character vector specifying the days to be selected (e.g. c("15", "16", "17")). If provided, only these days will be selected and the months and years adjusted (optional parameter).\cr}

\item{season}{A character string specifying the season to be selected (e.g. "DJF", "MAM", "JJA", "SON") (optional parameter) \cr}

\item{country_name}{A character string specifying the name of the country to be selected as provided by the 'rgeoboundaries' package.\cr
This is an optional parameter and can be used instead of the aoi parameter.}

\item{admin_level}{An integer specifying the administrative level of the country to be selected. Optional parameter than can be combined with the country_name parameter.}

\item{pressure_level}{A character string specifying the pressure level to be selected. This parameter is only relevant for pressure datasets (optional parameter).}

\item{target_file}{A character string specifying the name of the target file to store the downloaded data. \cr
If not provided, a default file name will be generated by oncatenating the variable name, StartDate and EndDate (optional parameter).}

\item{all_hours}{Logical. If TRUE, the function will download all hours of the day. If FALSE, the function will download only the time specified in the 'time' parameter. Default is FALSE (optional parameter).}

\item{path}{A character string specifying the path to store the downloaded data. If not provided, the data will be stored in the current working directory (optional parameter).}

\item{product_type}{A character string specifying the product type of the dataset. Some datasets require a product type to be specified (optional parameter).}

\item{grid}{A character string specifying the grid resolution of the dataset. The default is ".5/.5" for ERA5 datasets and ".1/.1" for ERA5 land datasets (optional parameter).}
}
\value{
A SpatRaster object containing the downloaded ERA5 data cropped to the extent of the aoi.7
}
\description{
This function allows users to download ERA5 data from the Copernicus Climate Data Store (CDS) API in R. \cr
The function uses the ecmwf package to access the CDS API and downloads the data in the NetCDF format. \cr
It will return the downloaded data as a SpatRaster object cropped to the extent of your aoi and can be used for further analysis. \cr
It simplifies the downloading of ERA5 datasets by summarizing the necessary steps to access the CDS API and providing filtering options . \cr
}
\examples{
# Download ERA5 data for temperature in Germany
user <- "Your user ID"
key <- " Your personal API key"
dataset_short_name <- "reanalysis-era5-land-monthly-means"
product_type <- "monthly_averaged_reanalysis"
variable <- "2m_temperature"
StartDate <- "2009-12-01"
EndDate <- "2020-02-29"
season <- "DJF"
time = sprintf("\%02d:00", 4:5) # c("04:00", "05:00")
country_name <- "Germany" # provided by the 'rgeoboundaries' package
data <- get_era5(user, key, dataset_short_name, variable, StartDate, EndDate, time, season=season, country_name=country_name, product_type=product_type)


}
