% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_extentt.R
\name{get_extent_rounded_2_grid}
\alias{get_extent_rounded_2_grid}
\title{Get the extent of any sf or terra object as input for the ERA5 function}
\usage{
get_extent_rounded_2_grid(spatial_obj, grid)
}
\arguments{
\item{spatial_obj}{The spatial object provided by the user in the ERA5 function}

\item{grid}{The grid size to which the extent should be rounded (e.g. "0.25/0.25")}
}
\value{
A character string with the extent in the format "North/West/South/East".
}
\description{
The function extracts the extent for any sf or terra object with rounded coordinates to the grid size specified by the user.\cr
The ERA5 data will be downloaded based on this extent.\cr
The function will return the extent in the desired format for the API: "North/West/South/East" (e.g., "51.75/7.25/51.5/7.5"). \cr
The function will also transform the spatial_object to the CRS of the ERA5 data (World Geodetic System 1984, EPSG: 4326) if it is not already in this CRS.
}
\examples{
spatial_obj <- st_read("Germany.shp")
grid <- "0.25/0.25"
get_extent_rounded_2_grid(spatial_obj, grid)
print(get_extent_rounded_2_grid)
"55.25/6/47.5/15.25"
}
