% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_Data.R
\name{get_seasonal_data}
\alias{get_seasonal_data}
\title{Get the months of a season as input for the ERA5 function}
\usage{
get_seasonal_data(season)
}
\arguments{
\item{season}{Character string. The season for which the months are required. Available inputs are "DJF", "MAM", "JJA", and "SON".}
}
\value{
A list of months for the specified season
}
\description{
This function returns the months of a season specified by the user. \cr
These months are used as input for the ERA5 function to download the data for the specified season.\cr
Available Seasons are: \cr
\itemize{
\item DJF ( December, January, February) \cr
\item MAM (March, April, May) \cr
\item JJA (June, July, August) \cr
\item SON (September, October, November)
}
}
\examples{
season <- "DJF"
months_season <- seasonal_data(season)
print(months_season)
[1] "12" "01" "02"

}
