% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours_aggregate.R
\name{hours_aggregate}
\alias{hours_aggregate}
\title{Aggregate hourly raster data to daily data
This function aggregates hourly raster data to daily data by calculating the mean or sum for each day.}
\usage{
hours_aggregate(data, aggr, areaAvg = FALSE)
}
\arguments{
\item{data}{Object of class SpatRaster.
Data must be hourly. If not all 24 hours of a day are provided, only the provided hours will be considered in the aggregation}

\item{aggr}{Method to aggregate the data. Can be mean() or sum().}

\item{areaAvg}{Logical. If TRUE, the function returns the daily sum() or mean() of the area average of the raster data. Default FALSE.}
}
\value{
Object of class SpatRaster with the daily mean or sum at grid-cell level or for the area average.
}
\description{
Aggregate hourly raster data to daily data
This function aggregates hourly raster data to daily data by calculating the mean or sum for each day.
}
