% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_plots.R
\name{monthly_plots}
\alias{monthly_plots}
\title{Plot analysis results for monthly raster data}
\usage{
monthly_plots(data, data_col, output_file)
}
\arguments{
\item{data}{Dataframe which contains statistics grouped by month. Monthly data can be analysed with raster_analysis.}

\item{data_col}{Column name of the data to plot.Can be mean, median, min, max, slope, mk_tau or mk_pval.}

\item{output_file}{Output file name. Should include the file extension.}
}
\value{
Plot showing selected data_col for all months at grid-cell level.
}
\description{
Plot analysis results for monthly raster data
}
