% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_aggregate.R
\name{months_aggregate}
\alias{months_aggregate}
\title{Aggregate monthly raster data to annual data
This function aggregates monthly raster data to annual data by calculating the mean or sum for each year.}
\usage{
months_aggregate(data, aggr, areaAvg = FALSE)
}
\arguments{
\item{data}{Object of class SpatRaster.
Data can be monthly or monthly averaged by hour of day with only one hour selected.
If target_months or seasons were used to download the data, only these months will be considered in the aggregation}

\item{aggr}{Method to aggregate the data. Can be mean() or sum().}

\item{areaAvg}{Logical. If TRUE, the function returns the annual sum() or mean() of the area average of the raster data. Default FALSE.}
}
\value{
Object of class SpatRaster with the annual mean or sum at grid-cell level or for the area average.
}
\description{
Aggregate monthly raster data to annual data
This function aggregates monthly raster data to annual data by calculating the mean or sum for each year.
}
