% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_data.R
\name{plot_annual_data}
\alias{plot_annual_data}
\title{Plot annual raster data}
\usage{
plot_annual_data(data, output_file)
}
\arguments{
\item{data}{Object of class SpatRaster which contains annual layers. Monthly data can be aggregated with months_aggregate().}

\item{output_file}{Output file name. Should include the file extension.}
}
\value{
Plot showing all years as raster in a grid.
}
\description{
Plot annual raster data
}
