% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{point_nc_to_df}
\alias{point_nc_to_df}
\title{Produces spatially weighted hourly data for a single location ready for use with runauto}
\usage{
point_nc_to_df(
  nc,
  x,
  y,
  start_time,
  end_time,
  lsm_nc = NULL,
  lsm_thresh = 0.05
)
}
\arguments{
\item{nc}{character vector containing nc filename(s). Data within nc files
must span the period defined by start_time and end_time.}

\item{x}{longitude of the location for which data is required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{y}{latitude of the location for which data is required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt object indicating the first hour for which data
is required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
is required.}

\item{lsm_nc}{character vector containing the ERA5 land/sea mask nc filename.}

\item{lsm_thresh}{threshold to divide land (0) and sea (1). Default = 0.05.}
}
\value{
a data frame containing hourly values for a suite of climate
variables.
}
\description{
`point_nc_to_df` takes an nc file or set of nc files containing
hourly ERA5 climate data, and for a given set of coordinates, produces an
inverse distance weighted mean of each variable, ready for use with
microclima::runauto.
}
\examples{

}
