% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_analysis.R
\name{raster_analysis}
\alias{raster_analysis}
\title{Monthly analysis at grid-cell level}
\usage{
raster_analysis(data)
}
\arguments{
\item{data}{Object of class SpatRaster.
Data can be monthly or monthly averaged by hour of day with only one hour selected.
If target_months or seasons were used to download the data, only these months will be considered in the analysis.}
}
\value{
A data frame with the following columns: x-coordinate, y-coordinate, months, mean, median, min, max, slope, mk_tau and mk_pval of each month over time at grid-cell level.
Data can be plotted with monthly_plots() from MCERA5.
}
\description{
This function groups input data by months and calculates statistics over time period of data set.
Calculated parameters are mean, median, min, max, Sen Slope, Mann-Kendall tau and Mann-Kendall p-value.
}
