% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{request_era5}
\alias{request_era5}
\title{Submit request(s) to CDS for ERA5 data download}
\usage{
request_era5(request, uid, out_path)
}
\arguments{
\item{request}{a list of request(s) created with `build_era5_request`.}

\item{uid}{character vector containing your CDS user ID.}

\item{out_path}{character vector of the location at which to download nc files.}
}
\description{
Uses the `ecwmfr` package to submit requests to the Climate
Data Store (CDS) for ERA5 .nc files. Executes one request at a time.
}
