% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclima_functions.R
\name{windheight}
\alias{windheight}
\title{Applies height correction to wind speed measurements}
\usage{
windheight(ui, zi, zo)
}
\arguments{
\item{ui}{numeric value(s) of measured wind speed (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{ m s^{-1}}}) at height `zi` (m).}

\item{zi}{a numeric value idicating the height (m) above the ground at which `ui` was measured.}

\item{zo}{a numeric value indicating the height (m) above ground level at which wind speed is desired.}
}
\value{
numeric values(s) of wind speed at height specified by `zo` (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{m s^{-1}}}).
}
\description{
`windheight` is used to to apply a height correction to wind speed measured at a specified height above ground level to obtain estimates of wind speed at a desired height above the ground.
}
\details{
Thus function assumes a logarithmic height profile to convert
wind speeds. It performs innacurately when `uo` is lower than 0.2
and a warning is given. If `uo` is below ~0.08 then the logairthmic height
profile cannot be used, and `uo` is converted to 0.1 and a warning given.
}
\examples{
windheight(3, 10, 1) # good
windheight(3, 10, 0.15) # performs poorly. Warning given
windheight(3, 10, 0.05) # cannot calculate. ui converted and warning given

}
\seealso{
The function [windcoef()] calculates a topographic or vegetation sheltering effect.
}
