% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_era5_request.R
\name{build_era5_request}
\alias{build_era5_request}
\title{Builds a request compatible with the CDS for all microclimate relevant climate
variables.}
\usage{
build_era5_request(
  xmin,
  xmax,
  ymin,
  ymax,
  start_time,
  end_time,
  by_month = TRUE,
  outfile_name = "era5_out"
)
}
\arguments{
\item{xmin}{The minimum longitude to request data for.}

\item{xmax}{The maximum longitude to request data for.}

\item{ymin}{The minimum latitude to request data for.}

\item{ymax}{The maximum latitude to request data for.}

\item{start_time}{a POSIXlt object indicating the first hour for which data
are required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
are required.}

\item{outfile_name}{character prefix for .nc files when downloaded (the year, month, and file extension is automatically added). Keep in
mind that you may want to submit multiple (or many) requests, and therefore
this prefix should adequately describe a unique query (e.g. by its
spatial or temporal extent).}
}
\description{
`build_era5_request` creates a request or set of requests that
can be submitted to the Climate Data Store (CDS) with the `ecmwfr` package.
Spatial and temporal extents are defined by the user, and requests are
automatically split by month. The following variables are requested:
2m_temperature, 2m_dewpoint_temperature, surface_pressure,
10m_u_component_of_wind, 10m_v_component_of_wind, total_precipitation,
total_cloud_cover, mean_surface_net_long_wave_radiation_flux,
mean_surface_downward_long_wave_radiation_flux,
total_sky_direct_solar_radiation_at_surface,
surface_solar_radiation_downwards.
}
