% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clim.R
\name{extract_clim}
\alias{extract_clim}
\title{Produces hourly data for a single location ready for use with
`microclima::runauto`.}
\usage{
extract_clim(
  nc,
  long,
  lat,
  start_time,
  end_time,
  d_weight = TRUE,
  dtr_cor = TRUE,
  dtr_cor_fac = 1
)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
`build_era5_request` and `request_era5` functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long}{longitude of the location for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat}{latitude of the location for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt object indicating the first hour for which data
are required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
are required.}

\item{d_weight}{logical value indicating whether to apply inverse distance
weighting using the 4 closest neighbouring points to the location defined by
`long` and `lat`. Default = `TRUE`.}

\item{dtr_cor}{logical value indicating whether to apply a diurnal temperature
range correction to air temperature values. Default = `TRUE`.}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction. Default = 1.}
}
\value{
a data frame containing hourly values for a suite of climate
variables.
}
\description{
`extract_clim` takes an nc file containing hourly ERA5 climate
data, and for a given set of coordinates, produces an (optionally) inverse
distance weighted mean of each variable, ready for use with
`microclima::runauto`. Also provides the option to implement a diurnal
temperature range correction to air temperatures.
}
\examples{

}
