% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_date_components.R
\name{extract_date_components}
\alias{extract_date_components}
\title{Extract date components as input for the ERA5 function}
\usage{
extract_date_components(
  StartDate,
  EndDate,
  dataset_short_name,
  year_interval = 1,
  month_interval = 1,
  day_interval = 1,
  target_years = NULL,
  target_days = NULL,
  seasonal_data = NULL,
  target_months = NULL
)
}
\arguments{
\item{StartDate}{Character string. The start date specified in the ERA5 function (e.g. "2000-01-01")}

\item{EndDate}{Character string. The end date specified in the ERA5 function (e.g. "2000-01-31")}

\item{dataset_short_name}{Character string. The short name of the dataset (e.g. "reanalysis-era5-land")}

\item{year_interval}{Integer. The interval for the years to be extracted. The default is 1 (optional parameter)}

\item{month_interval}{Integer. The interval for the months to be extracted. The default is 1 (optional parameter)}

\item{day_interval}{Integer. The interval for the days to be extracted. The default is 1 (optional parameter)}

\item{target_years}{Character vector. Can be used to only extract specific years (optional parameter).}

\item{target_days}{Character vector. Can be used to only extract specific days (optional parameter).}

\item{seasonal_data}{Character string. Can be used to extract the months of a specific season (optional parameter).\cr
For details, see the \code{\link{get_seasonal_data}} function.}

\item{target_months}{Character vector. Can be used to only extract specific months (optional parameter).}
}
\value{
A list with the extracted years, months and days
}
\description{
This function extracts the year, month and day between StartDate and EndDate specified in the ERA5 function. \cr
The API for the Copernicus Climate Data Store can only handle year, month and day as input for the download request. \cr
This function simplifies the process of extracting the year, month and day by allowing the user to specify a Start date and End date. \cr
Moreover, the user can specify an interval for the year, month and day, so that every n-th year, month and day is extracted. \cr
The user can also specify target years, months and days to extract only the specified years, months and days. \cr
Besides, the user can specify a season so that the function will extract the months of the specified season. \cr
If a monthly dataset is specified, only the year and month will be extracted.
}
\examples{
StartDate <- "2000-01-01"
EndDate <- "2000-01-31"
dataset_short_name <- "reanalysis-era5-land"
date_components <- extract_date_components(StartDate, EndDate, dataset_short_name)
print(date_components$days)
$years
[1] "2000"
$months
[1] "01"
$days
[1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10" "11" "12" "13" "14" "15" "16" "17" "18" "19" "20"
[21] "21" "22" "23" "24" "25" "26" "27" "28" "29" "30" "31"

Example using target_days
StartDate <- "2000-01-01"
EndDate <- "2000-12-31"
dataset_short_name <- "reanalysis-era5-land"
date_components <- extract_date_components(StartDate, EndDate, dataset_short_name, target_days = c("01", "15", "31"))
print(date_components)
$years
[1] "2000"
$months
[1] "01" "02" "03" "04" "05" "06" "07" "08" "09" "10" "11" "12"
$days
[1] "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15"
[21] "31" "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15" "31" "01" "15" "31"

Example using seasonal data
StartDate <- "2000-12-01"
EndDate <- "2002-02-01"
dataset_short_name <- "reanalysis-era5-land-monthly-means"
date_components <- extract_date_components(StartDate, EndDate, dataset_short_name, seasonal_data = "DJF")
print(date_components)
$years
[1] "2000" "2001" "2002"
$months
[1] "12" "01" "02" "12" "01" "02"
$days
NULL
}
