% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_precip.R
\name{extract_precip}
\alias{extract_precip}
\title{Produces daily precipitation data for a single location ready for use
with `microclima::runauto`.}
\usage{
extract_precip(nc, long, lat, start_time, end_time, d_weight = TRUE)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
`build_era5_request` and `request_era5` functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long}{longitude of the location for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat}{latitude of the location for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt object indicating the first hour for which data
are required.}

\item{end_time}{a POSIXlt object indicating the last hour for which data
are required.}

\item{d_weight}{logical value indicating whether to apply inverse distance
weighting using the 4 closest neighbouring points to the location defined by
`long` and `lat`. Default = `TRUE`.}
}
\value{
a numeric vector of daily precipitation (mm).
}
\description{
`extract_precip` takes an nc file containing hourly ERA5
climate data, and for a given set of coordinates, produces an (optionally)
inverse distance weighted mean of daily precipitation, ready for
use with `microclima::runauto`.
}
\examples{

}
