% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_data.R
\name{get_seasonal_data}
\alias{get_seasonal_data}
\title{Extract months for climatological seasons4}
\usage{
get_seasonal_data(season)
}
\arguments{
\item{season}{A character string or vector of climatological seasons. It can be one or more of the following: 'DJF', 'MAM', 'JJA', 'SON'}
}
\value{
A character vector of months corresponding to the climatological seasons
}
\description{
This function extracts the months for the climatological seasons that will be passed to the get_era5 function as target_months.\cr
The climatological seasons are defined as follows:\cr
\itemize{
\item DJF: December, January, February\cr
\item MAM: March, April, May\cr
\item JJA: June, July, August\cr
\item SON: September, October, November\cr
}
}
